Version parcial partida y repositorio [Cambiado algunos archivos Cuidado]

+++ src/main/java/org/springframework/samples/petclinic/usuario/UsuarioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepository extends CrudRepository<Usuario,Long> {
+public interface UsuarioRepository extends CrudRepository<Usuario,Integer> {
 	
 	public Collection<Usuario> findAll();
